+++ k8s.py
@@ -1,0 +1,7 @@
+import re
+
+
+def safe_name(name):
+    '''Replace "_", ".", "\", "/" in name by "-"'''
+    _sub = re.sub(r"_|\.|\\|/", "-", name)
+    return _sub.strip("-")
